Sequesters extraneous commands/synonyms in extra_commands.py.
Fixes #48
